Move shared memory dataset tests to own file
